refactor(config): break it apart
