+++ repo/config/datastore.go
@@ -1,0 +1,16 @@
+package config
+
+// DefaultDataStoreDirectory is the directory to store all the local IPFS data.
+const DefaultDataStoreDirectory = "datastore"
+
+// Datastore tracks the configuration of the datastore.
+type Datastore struct {
+	Type string
+	Path string
+}
+
+// DataStorePath returns the default data store path given a configuration root
+// (set an empty string to have the default configuration root)
+func DataStorePath(configroot string) (string, error) {
+	return Path(configroot, DefaultDataStoreDirectory)
+}
