+++ repo/config/mounts.go
@@ -1,0 +1,7 @@
+package config
+
+// Mounts stores the (string) mount points
+type Mounts struct {
+	IPFS string
+	IPNS string
+}
