+++ repo/config/tour.go
@@ -1,0 +1,7 @@
+package config
+
+// Tour stores the ipfs tour read-list and resume point
+type Tour struct {
+	Last string // last tour topic read
+	// Done []string // all topics done so far
+}
